### 2023-08-01 #5719 @editorialbot
[PRE REVIEW]: tightbinder: A Python package for semi-empirical tight-binding models of crystalline and disordered solids
**Submitting author:** @alejandrojuria (<a href="http://orcid.org/0000-0001-6668-7333">Alejandro José Uría-Álvarez</a>)
**Repository:** https://github.com/alejandrojuria/tightbinder
**Branch with paper.md** (empty if default branch): master
**Version:** v0.1
