### 2023-10-12 alejandrojuria/tightbinder#5 @yw-fang
[REVIEW comments] for the paper and code submitted to JOSS
Hello @alejandrojuria , thank you for your excellent work on this code.
Here are some comments that you may consider to improve the quality of the submitted manuscript (https://github.com/openjournals/joss-reviews/issues/5810) and the features of the code.

- [x]  The code, 'tightbinder', is reported to be an alternative tool that has some advantages over the popular tight binding code in the simulations of the realistic amorphous materials and the identification of the topological properties in the simulated systems. I didn't have experience with the other codes. Could you prepare a simple table or enumerated list to highlight the features of tightbinder that are not available or not easy to use in other codes? This highlight may be included in the README file. I know you have listed some functional features in the README, but let us know if these features are unique of tightbinder if you have already had the experience. I think it may help the upcoming new users to choose which codes they can take for their specific case.
- [x] In the manuscript, when mentioning the two Python classes designed for crystalline and amorphous solids, authors may include the specific names of the two classes.
- [x] On page 3 of the manuscript, change 'SK' to 'Slater-Koster' since it was only used once on line 83.
- [x] On line 95 of page 3, the authors are suggested to provide the website of the documentation which can be easily accessed by the readers.
- [x] I tested to build the documentation. I executed ``make html``, but it raised the error: ``make: *** No rule to make target 'html'.  Stop.
make: *** No rule to make target 'html'.  Stop.``
The "make'' refers to the version make-4.4.1 installed in my computer.
- [x] In all the tutorials, all paths for loading the files omit a dot (all my tests were run in a Linux mint system). However, if the paths are only shown for demo, it's okay and you don't need to change it.
- [x] In the tutorial for building the crystal https://tightbinder.readthedocs.io/en/latest/tutorial/crystal.html, execution of ``model.plot_wireframe()`` in Jupyter notebook raised an error ``TypeError: list indices must be integers or slices, not tuple''`` that seems to come from ``File ~/tightbinder/tightbinder/system.py:520, in System.plot_wireframe(self, ax, linewidth, alpha, color)''.
- [x] In the tutorial for setting up the device (https://tightbinder.readthedocs.io/en/latest/tutorial/transport.html), these example codes all raised parsing error ``ValueError: Filling must be a number``. I think there could be some errors in the examples/chain.txt.
- [x] Have you considered introducing ASE to this code (or Pymatgen or any other built-in methods in this code) so that the crystal can be built directly from the popular structure files in POSCAR and CIF? If it has been implemented, you're encouraged to mention it in the tutorial and the manuscript.
- [x] Automated tests are still lacking to verify the core functionalities (you can find below that these tests are required by Review Checklist of JOSS)

Here, I also linked this issue to the Review checklist https://github.com/openjournals/joss-reviews/issues/5810#issuecomment-1708561940 so that you can track the status of this review report.
