### 2023-10-30 d5ac032
Fix chain.txt file
Removed residual from previous commits
