### 2023-10-30 e6fd4fd
Motif stored as np.array
