### 2023-10-30 eb51154
Fix plot_wireframe tutorial
