### 2023-10-30 fe66c78
Fix visualization in transport device
