+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 sphinx
 pydata-sphinx-theme
 sphinx-autodoc-typehints
+sphinx_rtd_theme
