### 2023-11-07 cbf856c
Tests for SK class
