+++ tests/test_slaterkoster.py
@@ -1,0 +1,224 @@
+from tightbinder.models import SlaterKoster
+from tightbinder.fileparse import parse_config_file
+import numpy as np
+
+"""
+This file constains tests for both the System class and the SlaterKoster class, which inherits from System.
+The approach taken here is to test the SlaterKoster class, and then to test the System class by extension.
+These constitute functional tests of the library. 
+"""
+
+def test_configuration_parsing():
+    """
+    Function to test the parsing of a configuration file.
+    """
+
+    file = open("./examples/hBN.txt", "r")
+    config = parse_config_file(file)
+
+    assert config['Dimensionality'] == 2
+    
+    assert len(config['Bravais lattice']) == 2
+    assert config['Bravais lattice'][0] == [2.16506, 1.25, 0.0]
+    assert config['Bravais lattice'][1] == [2.16506, -1.25, 0.0]
+    
+    assert config['Species'] == 2
+    
+    assert len(config['Motif']) == 2
+    assert config['Motif'][0] == [0, 0, 0, 0]
+    assert config['Motif'][1] == [1.443376, 0, 0, 1]
+    
+    assert len(config['Filling']) == 2
+    assert config['Filling'][0] == 0.5
+    assert config['Filling'][1] == 0.5
+    
+    assert len(config['Orbitals']) == 2
+    assert config['Orbitals'][0] == ['s']
+    assert config['Orbitals'][1] == ['s']
+    
+    assert len(config['Onsite energy']) == 2
+    assert config['Onsite energy'][0] == [3.625]
+    assert config['Onsite energy'][0] == [3.625]
+    
+    assert len(config['SK amplitudes'].keys()) == 1
+    assert config['SK amplitudes']['1']['01'] == [-2.3, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+    
+    assert config['Spin'] == False 
+    
+    assert len(config['Spin-orbit coupling']) == 2
+    assert config['Spin-orbit coupling'][0] == 0
+    assert config['Spin-orbit coupling'][1] == 0
+    
+    assert len(config['Mesh']) == 2
+    assert config['Mesh'][0] == 100
+    assert config['Mesh'][1] == 100
+    
+    assert config['High symmetry points'] == ['G', "K", 'M', "G"]
+    
+
+def build_slaterkoster() -> SlaterKoster:
+    """
+    Function to build a SlaterKoster model from the configuration file.
+    """
+    
+    file = open("./examples/hBN.txt", "r")
+    config = parse_config_file(file)
+    
+    model = SlaterKoster(config)
+    
+    return model
+
+
+def test_sk_initialization():
+    
+    model = build_slaterkoster()
+    
+
+def test_supercell():
+    """
+    Routine to test the supercell method from System with a basic SlaterKoster model.
+    """
+    
+    model = build_slaterkoster().supercell(n1=2, n2=2)
+    
+    assert model.boundary == "PBC"
+    
+    assert model.natoms == 8    
+    assert np.allclose(model.bravais_lattice[0], [4.330128, 2.5, 0.0])
+    assert np.allclose(model.bravais_lattice[1], [4.330128, -2.5, 0.0])    
+    
+    expected_motif = np.array([[0, 0, 0, 0],
+                                [1.443376, 0, 0, 1],
+                                [2.16506, 1.25, 0.0, 0],
+                                [3.608436, 1.25, 0.0, 1],
+                                [2.16506, -1.25, 0.0, 0],
+                                [3.608436, -1.25, 0.0, 1],
+                                [4.330128, 0, 0, 0],
[…]
