### 2023-11-08 20cd915
Added more SK tests
