+++ tests/test_slaterkoster.py
@@ -1,11 +1,15 @@
-from tightbinder.models import SlaterKoster
+from tightbinder.models import SlaterKoster, AmorphousSlaterKoster
 from tightbinder.fileparse import parse_config_file
+from tightbinder.disorder import amorphize
 import numpy as np
+import math
+import matplotlib.pyplot as plt
 
 """
-This file constains tests for both the System class and the SlaterKoster class, which inherits from System.
-The approach taken here is to test the SlaterKoster class, and then to test the System class by extension.
-These constitute functional tests of the library. 
+This file constains tests for the main classes of the library, i.e. SlaterKoster and AmorphousSlaterKoster.
+This also serves to test the base classes System and Crystal, which are inherited by the former.
+All tests are functional, as they intend to check specific functionalities of the library and not 
+specific routines individually.
 """
 
 def test_configuration_parsing():
@@ -145,6 +149,19 @@ def test_add_bonds():
     assert model.bonds[2][3] == '2'
     
 
+def test_neighbour_distances():
+    """
+    Tests that neighbour distances are computed correctly.
+    """    
+    
+    model = build_slaterkoster().supercell(n1=4, n2=4)
+    model.initialize_hamiltonian()
+    
+    distances = model.compute_neighbour_distances(5)
+        
+    assert np.allclose(distances, [1.44, 2.5,  2.89, 3.82, 4.33])
+    
+
 def test_neighbours():
     """
     Tests the method find_neighbours from System.
@@ -221,4 +238,132 @@ def test_bloch_hamiltonian():
                        np.array([7.79426873,  6.85983334,  4.76896809,  3.625,       3.82600578,  4.1510816,
                         4.29309038,  5.38615122,  7.08312255,  7.79426873]))    
     
+
+def test_fermi_energy():
+    """
+    Tests that the Fermi energy is computed properly.
+    """
+    
+    model = build_slaterkoster()
+    model.initialize_hamiltonian()
+    
+    kpoints = model.high_symmetry_path(100, ['G', 'K', 'M', 'G'])
+    results = model.solve(kpoints)
+    
+    fermi_energy = results.calculate_fermi_energy(model.filling)
+    
+    assert math.isclose(fermi_energy, -3.625)
+    
+    
+def test_fermi_energy_supercell():
+    """
+    Tests that the Fermi energy is computed properly when working with supercells.
+    """
+    
+    model = build_slaterkoster().supercell(n1=2, n2=2)
+    model.initialize_hamiltonian()
+    
+    labels = ['G', 'K', 'M', 'G']
+    kpoints = model.high_symmetry_path(100, labels)
+    results = model.solve(kpoints)
+    
+    fermi_energy = results.calculate_fermi_energy(model.filling)
+    
+    assert math.isclose(fermi_energy, -3.625)
+    
+
+def test_rescale_bands_to_fermi_energy():
+    """
+    Tests method to set the Fermi energy to zero.
+    """
+    
+    model = build_slaterkoster()
+    model.initialize_hamiltonian()
+    
+    kpoints = model.high_symmetry_path(100, ['G', 'K', 'M', 'G'])
+    results = model.solve(kpoints)
+    
+    results.rescale_bands_to_fermi_level()
+    fermi_energy = results.calculate_fermi_energy(model.filling)
+    
+    assert math.isclose(fermi_energy, 0)
+    
+    
+def test_soc_bands():
+    """
+    Tests that SOC is correctly implemented.
+    """
+    
+    file = open("./examples/Bi111.txt", "r")
+    config = parse_config_file(file)
+    
+    model = SlaterKoster(config)
+    model.initialize_hamiltonian()
+    
[…]
