### 2023-11-08 6ee32be
Fixed bug in introduce_impurities
