### 2023-11-12 3bc7a6d
Added routine to hash arrays
