### 2023-11-12 550e088
Fix warning in dos_kpm
Also adjusted conductance output
