### 2023-11-12 4c7362e
Removed unnecesary lines
