### 2023-11-12 a2d2d92
Removed redundant Chern routine
Also fixed existing one
