+++ tests/test_slaterkoster.py
@@ -1,9 +1,11 @@
 from tightbinder.models import SlaterKoster, AmorphousSlaterKoster
 from tightbinder.fileparse import parse_config_file
 from tightbinder.disorder import amorphize
+from tightbinder.utils import hash_numpy_array
 import numpy as np
 import math
 import matplotlib.pyplot as plt
+
 
 """
 This file constains tests for the main classes of the library, i.e. SlaterKoster and AmorphousSlaterKoster.
@@ -89,18 +91,11 @@ def test_supercell():
     
     assert model.natoms == 8    
     assert np.allclose(model.bravais_lattice[0], [4.330128, 2.5, 0.0])
-    assert np.allclose(model.bravais_lattice[1], [4.330128, -2.5, 0.0])    
+    assert np.allclose(model.bravais_lattice[1], [4.330128, -2.5, 0.0])   
     
-    expected_motif = np.array([[0, 0, 0, 0],
-                                [1.443376, 0, 0, 1],
-                                [2.16506, 1.25, 0.0, 0],
-                                [3.608436, 1.25, 0.0, 1],
-                                [2.16506, -1.25, 0.0, 0],
-                                [3.608436, -1.25, 0.0, 1],
-                                [4.330128, 0, 0, 0],
-                                [5.773504, 0, 0, 1]])
+    motif_hash = hash_numpy_array(model.motif)
         
-    assert np.allclose(model.motif, expected_motif)
+    assert motif_hash == "7acd74da1ca11574db6c62a958d11a57"
 
 
 def test_reduce():
@@ -113,16 +108,9 @@ def test_reduce():
     assert model.boundary == "OBC"    
     assert model.natoms == 8    
     
-    expected_motif = np.array([[0, 0, 0, 0],
-                                [1.443376, 0, 0, 1],
-                                [2.16506, 1.25, 0.0, 0],
-                                [3.608436, 1.25, 0.0, 1],
-                                [2.16506, -1.25, 0.0, 0],
-                                [3.608436, -1.25, 0.0, 1],
-                                [4.330128, 0, 0, 0],
-                                [5.773504, 0, 0, 1]])
+    motif_hash = hash_numpy_array(model.motif)
         
-    assert np.allclose(model.motif, expected_motif)
+    assert motif_hash == "7acd74da1ca11574db6c62a958d11a57"
     
     
 def test_add_bonds():
@@ -207,17 +195,10 @@ def test_high_symmetry_path():
     model.initialize_hamiltonian()
     
     kpoints = model.high_symmetry_path(10, ['G', 'K', 'M', 'G'])
+    
+    kpoints_hash = hash_numpy_array(np.array(kpoints))
         
-    assert np.allclose(kpoints[0], np.array([0., 0., 0.]))
-    assert np.allclose(kpoints[1], np.array([0.        , 0.55850559, 0.        ])) 
-    assert np.allclose(kpoints[2], np.array([0.        , 1.11701117, 0.        ]))
-    assert np.allclose(kpoints[3], np.array([0.        , 1.67551676, 0.        ]))
-    assert np.allclose(kpoints[4], np.array([0.24184031, 1.53589019, 0.        ]))
-    assert np.allclose(kpoints[5], np.array([0.48368061, 1.39626363, 0.        ]))
-    assert np.allclose(kpoints[6], np.array([0.72552092, 1.25663706, 0.        ]))
-    assert np.allclose(kpoints[7], np.array([0.48368061, 0.83775804, 0.        ]))
-    assert np.allclose(kpoints[8], np.array([0.24184031, 0.41887902, 0.        ]))
-    assert np.allclose(kpoints[9], np.array([0., 0., 0.]))    
+    assert kpoints_hash == "30ae3de2af6b066896efe3148349fd1f"
[…]
