+++ tests/test_disorder.py
@@ -1,6 +1,7 @@
 from tightbinder.models import SlaterKoster
 from tightbinder.disorder import amorphize, introduce_vacancies, introduce_impurities
 from tightbinder.fileparse import parse_config_file
+from tightbinder.utils import hash_numpy_array
 import numpy as np
 
 """
@@ -20,18 +21,11 @@ def test_amorphize():
     
     model = SlaterKoster(config).supercell(n1=2, n2=2)
     model = amorphize(model, 0.2)
+    
+    motif_hash = hash_numpy_array(model.motif)
         
-    expected_motif = np.array([[-1.96545960e-01,  1.48990382e-01,  6.55557867e-02,  0.00000000e+00],
-                               [ 2.19673100e+00, -1.04642199e-01, -1.66581569e+00,  0.00000000e+00],
-                               [ 3.92584389e+00,  2.26661453e+00,  6.32944216e-05,  0.00000000e+00],
-                               [ 6.50041277e+00,  2.16755798e+00, -1.43467506e+00,  0.00000000e+00],
-                               [ 3.94072365e+00, -2.21612250e+00, -7.27792726e-02,  0.00000000e+00],
-                               [ 6.54716456e+00, -2.27049876e+00, -1.64122624e+00,  0.00000000e+00],
-                               [ 7.94529402e+00,  1.62597119e-02,  6.30498800e-02,  0.00000000e+00],
-                               [ 1.03855896e+01, -1.52747786e-01, -1.70513512e+00,  0.00000000e+00]])
-    
-    assert np.allclose(model.motif, expected_motif)
-    
+    assert motif_hash == "82e4913237df4fa0a84244c5857d443e"
+        
 
 def test_vacancies():
     """
@@ -45,13 +39,10 @@ def test_vacancies():
     
     model = SlaterKoster(config).supercell(n1=2, n2=2)
     model = introduce_vacancies(model, 0.2)
-            
-    expected_motif = np.array([[ 0.,       0.,       0.,       0.,    ],
-                               [ 2.61724,  0.,      -1.585,    0.     ],
-                               [ 6.54311,  2.2666,  -1.585,    0.     ],
-                               [10.46898,  0.,      -1.585,    0.     ]])
     
-    assert np.allclose(model.motif, expected_motif)
+    motif_hash = hash_numpy_array(model.motif)
+        
+    assert motif_hash == "1adc6630cb9136d79aa6efafe86a8387"
     
 
 def test_impurities():
@@ -68,13 +59,10 @@ def test_impurities():
     model.initialize_hamiltonian()
     
     model = introduce_impurities(model, 10, 0.9)
+    
+    hamiltonian_hash = hash_numpy_array(model.hamiltonian[0])
         
-    expected_hamiltonian = [[10. +0.j, -2.3+0.j,  0. +0.j,  0. +0.j],
-                            [-2.3+0.j, 10. +0.j, -2.3+0.j,  0. +0.j],
-                            [ 0. +0.j, -2.3+0.j, 10. +0.j, -2.3+0.j],
-                            [ 0. +0.j,  0. +0.j, -2.3+0.j, 10. +0.j]]
-                
-    assert np.allclose(model.hamiltonian[0], expected_hamiltonian)
+    assert hamiltonian_hash == "93609010a8c3a2f2ebae75938aa75e9a"
     
     
     
