### 2023-11-12 alejandrojuria/tightbinder#5 @alejandrojuria
Hello @yw-fang, thank you for your thorough review of the code. This past month I have been addressing the different issues and bugs you found, hopefully they are fixed now. Regarding some of your comments:

- As you say, the main purpose of the code is building Slater-Koster model, either crystalline or amorphous, and assess their topological nature. For crystalline systems, both are also done by other libraries, but as far as I know there are no libraries for building amorphous Slater-Koster models. I'm not experienced with the other codes, so the comparison is limited to what is listed on their respective websites. Instead of making a list, I have detailed more the features of the code in the README and highlighted the relevant parts, so hopefully the main characteristics come across faster.
- The instructions to build the documentation had an error, plus there was a missing dependency in the requirements.txt file. Those have been fixed now and the docs should build correctly.
- I was not familiar with ASE nor those formats you mention for describing structures. From what I read, I see that they are commonly used with other DFT codes. For this library, to build Slater-Koster models one has to specify the crystal structure together with the hopping amplitudes, and right now it cannot be done separately. However, it would make sense to be able to define the models this way, either defining the crystal algorithmically for example, or instead using those formats, and afterwards specifying the different hoppings. This is, the library does not support these formats currently but it would be reasonable to include them to facilitate interconnection with other codes and reach a wider audience.
- I have added several functional tests to the library, which can be run with pytest from the root folder of the project. These tests ensure that the functionality used for the example scripts works properly, although there are more tests than scripts to check each step required for the different types of calculations together with some core functionalities of the code.

Thank you for helping improve the code. If you have additional comments, we may discuss them here before closing the issue.
