### 2023-11-12 aa5810f
Refactored project structure
