### 2023-11-13 alejandrojuria/tightbinder#5 @yw-fang
Thank you for your responses! @alejandrojuria
I'll take close look at the details and get back to you soon (hopefully within this week)
