### 2023-11-13 3a9ae67
Added pyproject.toml file
Removed setup.py and refactored src folder
