+++ tightbinder/fileparse.py
@@ -1,511 +1,0 @@
-""" 
-Routines to parse the tight-binding parameters written in the input file.
-"""
-
-from io import TextIOWrapper
-import re
-from tightbinder.utils import pretty_print_dictionary
-from typing import List
-
-def parse_raw_arguments(file: TextIOWrapper) -> dict:
-    """ 
-    Routine to parse the arguments raw (without any treatment) from the input file, following
-    the rules defined for configuration files.
-
-    :param file: Pointer to configuration file, obtained from call to open().
-    :return: Dictionary with the content corresponding to each flag of the config. file. 
-    """
-
-    arguments = {}
-    value = []
-
-    lines = file.readlines()
-    for line in lines:
-        # Identifier
-        if line[0] == '#':
-            try:
-                arguments.update({new_argument: value})
-                value = []
-            except UnboundLocalError:
-                pass
-            new_argument = line[1:].strip()
-        # Blank line
-        elif line[0] in ['\n', '\r\n', '!', ' ']:
-            continue
-        # Argument values
-        else:
-            value.append(line.strip())
-    arguments.update({new_argument: value})
-
-    return arguments
-
-
-def check_arguments(arguments: dict, required_arguments: List[str]) -> None:
-    """ 
-    Routine to check whether all required fields for model are present in the input file.
-
-    :param arguments: Dictionary whose keys are the arguments to compare.
-    :param required_arguments: List with expected arguments.
-    :raises KeyError: Raise exception if not all required arguments are present in the arguments. 
-    """
-
-    for arg in required_arguments:
-        if arg not in arguments.keys():
-            raise KeyError(f"{arg} not present in input file")
-
-
-def shape_arguments(arguments: dict) -> dict:
-    """ Routine to rewrite correctly the arguments parsed (raw) from the input file.
-
-    :param arguments: Dictionary with the raw contents of the config. file.
-    :return: Dictionary with the contents of the config. file in numerical format. 
-    :raises IndexError: Raised if some arguments are missing
-    :raises ValueError: Raised if some arguments have incorrect values (e.g. string instead of numbers).
-    :raises SyntaxError: Raised for the SK amplitudes if not using correctly the brackets.
-    :raises NotImplementedError: Raised if there is an unexpected argument present. 
-    """
-
-    for arg in arguments:
-        if arg == 'System name':
-            try:
-                arguments[arg] = arguments[arg][0]
-            except IndexError as e:
-                print(f'{type(e).__name__}: No system name given')
-                raise
-
-        elif arg in ['Dimensionality', 'Species']:
-            try:
-                arguments[arg] = int(arguments[arg][0])
-            except IndexError as e:
-                print(f'{type(e).__name__}: No {arg} given')
-                raise
-            except ValueError as e:
-                print(f'{type(e).__name__}: {arg} has to be an integer')
-                raise
-
-        elif arg == 'Bravais lattice':
-            aux_array = []
-            for line in arguments[arg]:
-                try:
-                    aux_array.append([float(num) for num in re.split(' |, |,', line)])
-                except IndexError as e:
-                    print(f'{type(e).__name__}: No {arg} vectors given given')
-                    raise
-                except ValueError as e:
-                    print(f'{type(e).__name__}: {arg} vectors have to be numbers')
-                    raise
-            arguments[arg] = aux_array
-
-        elif arg == 'Motif':
-            aux_array = []
-            for n, line in enumerate(arguments[arg]):
-                aux_array.append([float(num) for num in re.split(' |, |,|; |;', line)])
-                if arguments['Species'] == 1:
-                    try:
-                        aux_array[n][3] = 0  # Default value to 1
[…]
