+++ tightbinder/observables.py
@@ -1,937 +1,0 @@
-"""
-Module containing routines for computation of observables.
-"""
- 
-from tightbinder.result import Spectrum
-from tightbinder.system import System
-from tightbinder.models import SlaterKoster
-
-import numpy as np
-import scipy.sparse as sp
-import matplotlib.pyplot as plt
-from matplotlib.axes import Axes
-from typing import Union, List, Tuple
-from copy import deepcopy
-
-
-def _retarded_green_function(w: float, e: float, delta: float) -> complex:
-    """ 
-    Routine to compute the retarded Green function, to be used to obtain
-    the DOS.
-
-    :param w: Value of frequency.
-    :param e: Value of energy.
-    :param delta: Value of broadening.
-    :return: Retarded Green's function evaluated at w, e. 
-    """
-    
-    green = 1./(w - e + 1j*delta)
-    return green
-
-
-def dos(result: Spectrum, energy: float = None, delta: float = 0.1, npoints: int = 200) -> Tuple[List[float], List[float]]:
-    """ 
-    Routine to compute the density of states from a Result object.
-
-    :param result: Result object.
-    :param energy: Value of energy where DoS is computed. If None, then the DoS is computed
-        for the whole energy window of the spectrum.
-    :param delta: Delta broadening, defaults to 0.01. 
-    :param npoints: Number of energy points to use if sampling whole DoS.
-    :return: Returns duple with values of dos, and energies where was computed.
-    """
-    
-    nkpoints = result.eigen_energy.shape[1]
-    eigval = result.eigen_energy.reshape(-1, )
-    if not energy:
-        emin, emax = np.min(result.eigen_energy)*1.5, np.max(result.eigen_energy)*1.5
-        energies = np.linspace(emin, emax, npoints, endpoint=True)
-    else:
-        energies = [energy]
-
-    energies = np.array(energies)
-    dos = []
-    for energy in energies:
-        density = (np.sum(-np.imag(_retarded_green_function(energy, eigval, delta))/np.pi)
-                /(nkpoints))
-        dos.append(density)
-
-    return dos, energies
-
-
-def dos_kpm(system: System, energy: float = None, npoints: int = 200, nmoments: int = 30, r: int = 10) -> Tuple[List[float], List[float]]:
-    """ 
-    Routine to compute the density of states using the Kernel Polynomial Method.
-    Intended to be used with supercells and k=0.
-
-    :param system: System whose density of states we want to compute.
-    :param energy: Value of energy where we want to obtain the DoS. If None, DoS is computed for all
-        energy window of spectrum.
-    :param npoints: Number of energy points to consider if computing full DoS. Defaults to 200.
-    :param nmoments: Number of moments to use in the KPM expansion. Defaults to 30.
-    :param r: Number of samples to use in the stochastic evaluation of the KPM traces.
-        Defaults to 10.
-    :return:
-    """
-
-    if system.matrix_type != "sparse":
-        print("Warning: KPM computations are intended to be run with sparse matrices (faster and less memory usage)")
-
-    print("Computing DOS using the KPM...")
-    h = system.hamiltonian_k([0., 0., 0.])
-    if system.matrix_type == "dense":
-        h = sp.bsr_matrix(h)
-    # h spectrum has to be between -1 and 1
-    h_norm = sp.linalg.norm(h, ord=np.inf)
-    h /= h_norm
-    
-    if not energy:
-        emin, emax = -1, 1
-        energies = np.linspace(emin, emax, npoints, endpoint=True)[1:-1]
-    else:
-        energies = [energy/h_norm]
-
-    energies = np.array(energies)
-    moments = compute_dos_momentum(nmoments, h, r)
-    jackson = jackson_kernel(nmoments)
[…]
