+++ setup.py
@@ -1,6 +1,0 @@
-from distutils.core import setup
-
-setup(name='tightbinder',
-      version='0.1',
-      description='Slater-Koster tight-binding library for crystalline and disordered solids',
-      packages=['tightbinder',])
