### 2023-11-13 alejandrojuria/tightbinder#5 @yw-fang
@alejandrojuria

Here is an additional comment I can make based on your latest version 3a9ae67a224e4cbdf9011b65fdfac239e998501a:

`make html ` now works after you fixed it. But I noticed that it requires several depdendencies installed. If you dislike including them in the requirements.txt, you'd better indicate them in the README.md file.
In my case, I found I have to install these libraries via pip:
pip install sphinx_rtd_theme
pip install sphinx-autodoc-typehints
pip install pydata_sphinx_theme
