### 2023-11-13 45a2fe9
Changed hash function
Also adjusted all tests accordingly
