+++ tests/test_disorder.py
@@ -3,6 +3,8 @@ from tightbinder.disorder import amorphize, introduce_vacancies, introduce_impur
 from tightbinder.fileparse import parse_config_file
 from tightbinder.utils import hash_numpy_array
 import numpy as np
+import math
+
 
 """
 Tests to check functionality of the disorder module.
@@ -24,7 +26,7 @@ def test_amorphize():
     
     motif_hash = hash_numpy_array(model.motif)
         
-    assert motif_hash == "82e4913237df4fa0a84244c5857d443e"
+    assert math.isclose(motif_hash, 322.32157833573854)
         
 
 def test_vacancies():
@@ -42,7 +44,7 @@ def test_vacancies():
     
     motif_hash = hash_numpy_array(model.motif)
         
-    assert motif_hash == "1adc6630cb9136d79aa6efafe86a8387"
+    assert math.isclose(motif_hash, 171.9359264901)
     
 
 def test_impurities():
@@ -60,9 +62,9 @@ def test_impurities():
     
     model = introduce_impurities(model, 10, 0.9)
     
-    hamiltonian_hash = hash_numpy_array(model.hamiltonian[0])
+    hamiltonian_hash = np.real(hash_numpy_array(model.hamiltonian[0]))
         
-    assert hamiltonian_hash == "93609010a8c3a2f2ebae75938aa75e9a"
+    assert math.isclose(hamiltonian_hash, 431.73999999999995)
     
     
     
