+++ .readthedocs.yaml
@@ -32,5 +32,7 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
+    - method: pip
+      path: .
     - requirements: docs/requirements.txt
-    - requirements: requirements.txt
+
