### 2023-11-15 c63afd0
improved file finding in tests
