### 2023-11-15 7cef7e8
Specify optional deps in pyproject.toml
