+++ .readthedocs.yaml
@@ -34,5 +34,6 @@ python:
   install:
     - method: pip
       path: .
-    - requirements: docs/requirements.txt
+    - method: pip
+      path: .[docs]
 
