### 2023-11-15 alejandrojuria/tightbinder#12 @mbkumar
Archaic text based input file format
I looked at the input file format and the code used to parse the file. Both are, to be honest, not well implemented. Here are some thoughts on this.

1. As the other reviewer mentioned, you could have used well-known formats such as ASE, pymatgen or VASP formats for the input crystal structure.
2. The species type is missing in all the example input files.
3.  If you don't want to introduce big packages such as pymatgen as dependencies, the input file is essentially a dictionary/map. You could have used YAML or JSON for this. Those formats give you all the flexibility and you don't have write custom code to parse the text. For example GaSe.txt could be GaSe.yaml as follows
```yaml
SystemName: GaSe
Dimensions: 2
Lattice:
  - 3.30735 -1.90950  0.00000
  - 0.00000  3.81900  0.00000
Species: [Ga, Se] # Instead of 2, I am putting the species types
Motif: [
  -1.102  -1.910  1.235  0
  -1.102  -1.909 -1.235  0
  -2.205  -0.000  2.409  1
  ]
Filling: [0.5, 0.5]
Orbitals: [pz, pz]
OnsiteEnergy: [0.41, -0.41]
# OnSiteEnergyUnit, if you want to specify different units
SKAmplitudes:
  - [0, 1; 1] 0 -2.33 # This line may not be parsed properly by yaml, but could be easily reformatted for yaml
  - [0, 1; 2] 0 0.61
  - [0, 0; 3] 0 0.07
  - [1, 1, 3] 0 0.07
  - [0, 1; 4] 0 0.13
  - [0, 0; 5] 0 -0.09
  - [1, 1, 5] 0 -0.09
Spin: False
SOC: [0, 0]
Mesh: [100, 100]
CriticalPoints: M
```
and the code to parse the file would be very simple as below.
```python
import yaml

# Read YAML file
with open("GaSe.yaml") as fp:
    data = yaml.safe_load(fp)
    configuration = shape_arguments(data)
```
