### 2023-11-15 alejandrojuria/tightbinder#12 @alejandrojuria
It is true that the format I have defined is very basic. At the moment of writing I was not aware of these formats, as you say it might be better to simply refactor the code and use the yaml format. I will test the format (specially the part about SK amplitudes); if the transition is simple I will do the switch.
