### 2023-11-15 alejandrojuria/tightbinder#12 @mbkumar
Yes. Read up on the YAML format. It is easy on human eyes as well as machine readable. Should not be that difficult to make the switch. In the mean while, I will check the other parts of the code.
