### 2023-11-15 677243e
Execution of tests from arbitrary folders
Also changed paths to reflect new project structure
