+++ tests/test_observables.py
@@ -4,6 +4,7 @@ from tightbinder.models import SlaterKoster
 from tightbinder.utils import hash_numpy_array
 import numpy as np
 import math
+from pathlib import Path
 
 """
 Module with tests to verify the functionality of the observables module.
@@ -15,21 +16,22 @@ def test_dos():
     Tests calculation of the density of states.
     """
     
-    file = open("./examples/hBN.txt", "r")
-    config = parse_config_file(file)
-    model = SlaterKoster(config)
+    path = Path(__file__).parent / ".." / "examples" / "inputs" / "hBN.txt"
+    with open(path) as fp:
+        config = parse_config_file(fp)
+        model = SlaterKoster(config)
 
-    nk = 20
-    kpoints = model.brillouin_zone_mesh([nk, nk])
+        nk = 20
+        kpoints = model.brillouin_zone_mesh([nk, nk])
 
-    model.initialize_hamiltonian()
-    results = model.solve(kpoints)
+        model.initialize_hamiltonian()
+        results = model.solve(kpoints)
 
-    density, energies = dos(results, delta=0.05, npoints=200)
-    
-    density_hash = hash_numpy_array(np.array(density))
-    
-    assert math.isclose(density_hash, 5.678625492981987)
+        density, energies = dos(results, delta=0.05, npoints=200)
+        
+        density_hash = hash_numpy_array(np.array(density))
+        
+        assert math.isclose(density_hash, 5.678625492981987)
             
 
 def test_dos_kpm():
@@ -39,20 +41,21 @@ def test_dos_kpm():
     
     np.random.seed(1)
 
-    file = open("./examples/hBN.txt", "r")
-    config = parse_config_file(file)
+    path = Path(__file__).parent / ".." / "examples" / "inputs" / "hBN.txt"
+    with open(path) as fp:
+        config = parse_config_file(fp)
 
-    ncells = 10
-    model = SlaterKoster(config).supercell(n1=ncells, n2=ncells)
+        ncells = 10
+        model = SlaterKoster(config).supercell(n1=ncells, n2=ncells)
 
-    model.matrix_type = "sparse"
-    model.initialize_hamiltonian()
+        model.matrix_type = "sparse"
+        model.initialize_hamiltonian()
 
-    density, energies = dos_kpm(model, nmoments=150, npoints=400, r=30)
-    
-    density_hash = hash_numpy_array(np.array(density))
-    
-    assert math.isclose(density_hash, 0.7510637669286024)
+        density, energies = dos_kpm(model, nmoments=150, npoints=400, r=30)
+        
+        density_hash = hash_numpy_array(np.array(density))
+        
+        assert math.isclose(density_hash, 0.7510637669286024)
             
 
 def test_transport_device():
@@ -62,26 +65,27 @@ def test_transport_device():
     
     length, width = 10, 5
 
-    file = open("examples/chain.txt", "r")
-    config = parse_config_file(file)
+    path = Path(__file__).parent / ".." / "examples" / "inputs" / "chain.txt"
+    with open(path) as fp:
+        config = parse_config_file(fp)
 
-    model = SlaterKoster(config)
+        model = SlaterKoster(config)
 
-    model.bravais_lattice = np.concatenate((model.bravais_lattice, np.array([[0., 1, 0]])))
-    model = model.reduce(n2=width)
+        model.bravais_lattice = np.concatenate((model.bravais_lattice, np.array([[0., 1, 0]])))
+        model = model.reduce(n2=width)
 
-    left_lead = np.copy(model.motif)
-    left_lead[:, :3] -= model.bravais_lattice[0]
+        left_lead = np.copy(model.motif)
+        left_lead[:, :3] -= model.bravais_lattice[0]
 
[…]
