### 2023-11-15 164e2e2
Remove requirements file
Update docs yaml
