+++ requirements.txt
@@ -1,4 +1,0 @@
-numpy
-scipy
-matplotlib
-vpython
