+++ .readthedocs.yaml
@@ -32,8 +32,5 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
-    - method: pip
-      path: .
-    - method: pip
       path: .[docs]
 
