+++ .readthedocs.yaml
@@ -32,5 +32,6 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
+    - method: pip
       path: .[docs]
 
