### 2023-11-15 96c620e
Use INPUT_FILE_DIR in tests and examples for specifying input files
