+++ tests/test_disorder.py
@@ -6,7 +6,7 @@ import numpy as np
 import math
 from pathlib import Path
 
-
+INPUT_FILE_DIR = Path(__file__).parent / ".." / "examples" / "inputs"
 """
 Tests to check functionality of the disorder module.
 """
@@ -19,7 +19,7 @@ def test_amorphize():
     
     np.random.seed(1)
     
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.txt"
+    path = INPUT_FILE_DIR / "Bi111.txt"
     with open(path) as fp:
         config = parse_config_file(fp)
     
@@ -38,7 +38,7 @@ def test_vacancies():
     
     np.random.seed(1)
     
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.txt"
+    path = INPUT_FILE_DIR / "Bi111.txt"
     with open(path) as fp:
         config = parse_config_file(fp)
     
@@ -57,7 +57,7 @@ def test_impurities():
     
     np.random.seed(1)
     
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "hBN.txt"
+    path = INPUT_FILE_DIR / "hBN.txt"
     with open(path) as fp:
         config = parse_config_file(fp)
     
