+++ examples/inputs/GaSe.txt
@@ -1,82 +1,0 @@
-! --------------------- Config file formatting rules ---------------------
-! Lines starting with ! are comments
-! Lines starting with # denote argument types and must be present (removing one or wirting it wrogn will result in error)
-! Blank spaces are irrelevant
-! All numbers must be separated by either spaces, commas or semicolons
-! BEWARE: Leaving a blank space at the beginning of a line will result in it being skipped
-!
-! This model is based on the paper "Excitons in Boron Nitride Single Layer"
-! --------------------- Material parameters ---------------------
-#System name
-GaSe
-
-#Dimensionality
-2
-
-#Bravais lattice
-! Each row corresponds to one vector of the basis (vectors have to be in R^3)
-! For 1D and 2D systems, the z (third) component must be zero (z should always be regarded as height component)
-3.30735 -1.90950  0.00000
-0.00000  3.81900  0.00000
-
-#Species
-! Number of atomic species (isoelectronic at the moment)
-2
-
-#Motif
-! Each row corresponds to one vector (atom) of the basis; fourth number denotes atomic species
-! Species have to be either 0 or 1
--1.102  -1.910  1.235  0
--1.102  -1.909 -1.235  0
--2.205  -0.000  2.409  1
-
-
-#Filling
-! Must provide filling per species, each filling can be fractional if the system is spinless.
-! Sum of the total number of electrons in the system must be an integer.
-0.5
-0.5
-
-#Orbitals
-! Always use cubic harmonic orbitals
-! Order is irrelevant
-! Species have to be isoelectronic (same number of valence shell electrons)
-! Possible orbitals are: s, px, py, pz, dxy, dyz, dzx, dx2-y2, d3z2-r2
-pz
-pz
-
-#Onsite energy
-! One for each different orbital defined above
-! In case of having two atomic species, two lines must be appended, one for each element
-0.41
--0.41
-
-#SK amplitudes
-! Order is fixed: Vsss, Vsps, Vpps, Vppp, Vsds, Vpds, Vpdp, Vdds, Vddp, Vddd (separated by spaces)
-! Not present amplitudes can be omitted but ordering has te be respected always
-! In case of having two atomic species, two lines must be appended, one for each element
-[0, 1; 1] 0 -2.33
-[0, 1; 2] 0 0.61
-[0, 0; 3] 0 0.07
-[1, 1, 3] 0 0.07
-[0, 1; 4] 0 0.13
-[0, 0; 5] 0 -0.09
-[1, 1, 5] 0 -0.09
-
-#Spin
-! Determine whether the model is spinless (spin polarized) or spinful
-! Must be either True (spinful) or False (spinless); if left blank it will default to False
-! NOTE: True or False must be capitalized
-False
-
-#Spin-orbit coupling
-0
-0
-
-! --------------------- Simulation parameters ---------------------
-#Mesh
-100 100
-
-#High symmetry points
-! Label of points which make the path to evalute the bands of the system
-M 
