### 2023-11-16 7de104e
New input file format
Now using yaml instead of the previous, custom one
