+++ src/tightbinder/fileparse.py
@@ -6,6 +6,8 @@ from io import TextIOWrapper
 import re
 from tightbinder.utils import pretty_print_dictionary
 from typing import List
+import yaml
+from numbers import Number
 
 def parse_raw_arguments(file: TextIOWrapper) -> dict:
     """ 
@@ -66,46 +68,23 @@ def shape_arguments(arguments: dict) -> dict:
     """
 
     for arg in arguments:
-        if arg == 'System name':
-            try:
-                arguments[arg] = arguments[arg][0]
-            except IndexError as e:
-                print(f'{type(e).__name__}: No system name given')
-                raise
-
-        elif arg in ['Dimensionality', 'Species']:
-            try:
-                arguments[arg] = int(arguments[arg][0])
-            except IndexError as e:
-                print(f'{type(e).__name__}: No {arg} given')
-                raise
-            except ValueError as e:
-                print(f'{type(e).__name__}: {arg} has to be an integer')
-                raise
-
-        elif arg == 'Bravais lattice':
-            aux_array = []
-            for line in arguments[arg]:
-                try:
-                    aux_array.append([float(num) for num in re.split(' |, |,', line)])
-                except IndexError as e:
-                    print(f'{type(e).__name__}: No {arg} vectors given given')
-                    raise
-                except ValueError as e:
-                    print(f'{type(e).__name__}: {arg} vectors have to be numbers')
-                    raise
-            arguments[arg] = aux_array
-
-        elif arg == 'Motif':
-            aux_array = []
+        
+        if arg == 'Species':
+            if type(arguments[arg]) is str:
+                arguments[arg] = [arguments[arg]]
+        
+        elif arg in ['Filling', 'SOC']:
+            if isinstance(arguments[arg], Number):
+                arguments[arg] = [float(arguments[arg])]
+                
+        elif arg == 'Mesh':
+            if isinstance(arguments[arg], Number):
+                arguments[arg] = [int(arguments[arg])]
+                
+        elif arg == 'OnsiteEnergy':
             for n, line in enumerate(arguments[arg]):
-                aux_array.append([float(num) for num in re.split(' |, |,|; |;', line)])
-                if arguments['Species'] == 1:
-                    try:
-                        aux_array[n][3] = 0  # Default value to 1
-                    except IndexError:
-                        aux_array[n].append(0)
-            arguments[arg] = aux_array
+                if isinstance(line, Number):
+                    arguments[arg][n] = [float(line)] 
 
         elif arg == 'Orbitals':
             possible_orbitals = ['s', 'px', 'py', 'pz', 'dxy', 'dyz', 'dzx', 'dx2-y2', 'd3z2-r2']
@@ -126,31 +105,18 @@ def shape_arguments(arguments: dict) -> dict:
                     aux_array.append(orbitals)
             arguments[arg] = aux_array
 
-        elif arg == 'Onsite energy':
-            aux_array = []
-            for line in arguments[arg]:
-                try:
-                    aux_array.append([float(num) for num in re.split(' |, |,', line)])
-                except IndexError as e:
-                    print(f'{type(e).__name__}: No onsite energies included')
-                    raise
-                except ValueError as e:
-                    print(f'{type(e).__name__}: Onsite energies must be numbers')
-                    raise
-            arguments[arg] = aux_array
-
-        elif arg == 'SK amplitudes':
+        elif arg == 'SKAmplitudes':
             dictionary = {}
             for n, line in enumerate(arguments[arg]):
                 if not line:
                     raise SyntaxError('No Slater-Koster amplitudes given')
                 else:
                     # First try to parse square brakets
[…]
