+++ tests/test_topology.py
@@ -7,6 +7,7 @@ import matplotlib.pyplot as plt
 import math
 from pathlib import Path
 
+INPUT_FILE_DIR = Path(__file__).parent / ".." / "examples" / "inputs"
 """
 Module with tests to check the topology module.
 """
@@ -16,18 +17,17 @@ def test_wannier_centre_flow():
     Tests the calculation of the wannier centre flow for a SlaterKoster model.
     """
     
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "Bi111.yaml"
+    config = parse_config_file(path)
+    
+    model = SlaterKoster(config)
+    model.initialize_hamiltonian()
+    
+    wcc = calculate_wannier_centre_flow(model, 5, refine_mesh=False)
+    
+    wcc_hash = hash_numpy_array(wcc)
         
-        model = SlaterKoster(config)
-        model.initialize_hamiltonian()
-        
-        wcc = calculate_wannier_centre_flow(model, 5, refine_mesh=False)
-        
-        wcc_hash = hash_numpy_array(wcc)
-            
-        assert math.isclose(wcc_hash, 9.241406000000001)
+    assert math.isclose(wcc_hash, 9.241406000000001)
     
 
 def test_wannier_centre_flow_refined():
@@ -35,18 +35,17 @@ def test_wannier_centre_flow_refined():
     Tests the calculation of the wannier centre flow but with a refined mesh.
     """
     
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "Bi111.yaml"
+    config = parse_config_file(path)
         
-        model = SlaterKoster(config)
-        model.initialize_hamiltonian()
-        
-        wcc = calculate_wannier_centre_flow(model, 5)
-        
-        wcc_hash = hash_numpy_array(wcc)
-        
-        assert math.isclose(wcc_hash, 22.862526000000003)
+    model = SlaterKoster(config)
+    model.initialize_hamiltonian()
+    
+    wcc = calculate_wannier_centre_flow(model, 5)
+    
+    wcc_hash = hash_numpy_array(wcc)
+    
+    assert math.isclose(wcc_hash, 22.862526000000003)
             
 
 def test_z2_invariant():
@@ -54,17 +53,16 @@ def test_z2_invariant():
     Tests the z2 invariant calculation algorithm.
     """
     
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "Bi111.yaml"
+    config = parse_config_file(path)
     
-        model = SlaterKoster(config)
-        model.initialize_hamiltonian()
-        
-        wcc = calculate_wannier_centre_flow(model, 10, refine_mesh=False)
-        z2 = calculate_z2_invariant(wcc)
-        
-        assert z2 == 1
+    model = SlaterKoster(config)
+    model.initialize_hamiltonian()
+    
+    wcc = calculate_wannier_centre_flow(model, 10, refine_mesh=False)
+    z2 = calculate_z2_invariant(wcc)
+    
+    assert z2 == 1
     
 
 def test_z2_phase_diagram():
@@ -72,25 +70,24 @@ def test_z2_phase_diagram():
     Tests that the z2 invariant calculation algorithm works to determine a phase diagram, changing the spin-orbit coupling.
     """
     
[…]
