+++ tests/test_slaterkoster.py
@@ -4,10 +4,9 @@ from tightbinder.disorder import amorphize
 from tightbinder.utils import hash_numpy_array
 import numpy as np
 import math
-import matplotlib.pyplot as plt
 from pathlib import Path
 
-
+INPUT_FILE_DIR = Path(__file__).parent / ".." / "examples" / "inputs"
 """
 This file constains tests for the main classes of the library, i.e. SlaterKoster and AmorphousSlaterKoster.
 This also serves to test the base classes System and Crystal, which are inherited by the former.
@@ -20,48 +19,47 @@ def test_configuration_parsing():
     Function to test the parsing of a configuration file.
     """
 
-    path = Path(__file__).parent / ".." / "examples" / "inputs" / "hBN.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "hBN.yaml"
+    config = parse_config_file(path)
 
-        assert config['Dimensionality'] == 2
-        
-        assert len(config['Bravais lattice']) == 2
-        assert config['Bravais lattice'][0] == [2.16506, 1.25, 0.0]
-        assert config['Bravais lattice'][1] == [2.16506, -1.25, 0.0]
-        
-        assert config['Species'] == 2
-        
-        assert len(config['Motif']) == 2
-        assert config['Motif'][0] == [0, 0, 0, 0]
-        assert config['Motif'][1] == [1.443376, 0, 0, 1]
-        
-        assert len(config['Filling']) == 2
-        assert config['Filling'][0] == 0.5
-        assert config['Filling'][1] == 0.5
-        
-        assert len(config['Orbitals']) == 2
-        assert config['Orbitals'][0] == ['s']
-        assert config['Orbitals'][1] == ['s']
-        
-        assert len(config['Onsite energy']) == 2
-        assert config['Onsite energy'][0] == [3.625]
-        assert config['Onsite energy'][0] == [3.625]
-        
-        assert len(config['SK amplitudes'].keys()) == 1
-        assert config['SK amplitudes']['1']['01'] == [-2.3, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-        
-        assert config['Spin'] == False 
-        
-        assert len(config['Spin-orbit coupling']) == 2
-        assert config['Spin-orbit coupling'][0] == 0
-        assert config['Spin-orbit coupling'][1] == 0
-        
-        assert len(config['Mesh']) == 2
-        assert config['Mesh'][0] == 100
-        assert config['Mesh'][1] == 100
-        
-        assert config['High symmetry points'] == ['G', "K", 'M', "G"]
+    assert config['Dimensions'] == 2
+    
+    assert len(config['Lattice']) == 2
+    assert config['Lattice'][0] == [2.16506, 1.25, 0.0]
+    assert config['Lattice'][1] == [2.16506, -1.25, 0.0]
+    
+    assert len(config['Species']) == 2
+    
+    assert len(config['Motif']) == 2
+    assert config['Motif'][0] == [0, 0, 0, 0]
+    assert config['Motif'][1] == [1.443376, 0, 0, 1]
+    
+    assert len(config['Filling']) == 2
+    assert config['Filling'][0] == 0.5
+    assert config['Filling'][1] == 0.5
+    
+    assert len(config['Orbitals']) == 2
+    assert config['Orbitals'][0] == ['s']
+    assert config['Orbitals'][1] == ['s']
+    
+    assert len(config['OnsiteEnergy']) == 2
+    assert config['OnsiteEnergy'][0] == [3.625]
+    assert config['OnsiteEnergy'][0] == [3.625]
+    
[…]
