+++ tests/test_disorder.py
@@ -19,16 +19,15 @@ def test_amorphize():
     
     np.random.seed(1)
     
-    path = INPUT_FILE_DIR / "Bi111.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "Bi111.yaml"
+    config = parse_config_file(path)
+
+    model = SlaterKoster(config).supercell(n1=2, n2=2)
+    model = amorphize(model, 0.2)
+
+    motif_hash = hash_numpy_array(model.motif)
     
-        model = SlaterKoster(config).supercell(n1=2, n2=2)
-        model = amorphize(model, 0.2)
-    
-        motif_hash = hash_numpy_array(model.motif)
-        
-        assert math.isclose(motif_hash, 322.32157833573854)
+    assert math.isclose(motif_hash, 322.32157833573854)
         
 
 def test_vacancies():
@@ -38,16 +37,15 @@ def test_vacancies():
     
     np.random.seed(1)
     
-    path = INPUT_FILE_DIR / "Bi111.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "Bi111.yaml"
+    config = parse_config_file(path)
+
+    model = SlaterKoster(config).supercell(n1=2, n2=2)
+    model = introduce_vacancies(model, 0.2)
+
+    motif_hash = hash_numpy_array(model.motif)
     
-        model = SlaterKoster(config).supercell(n1=2, n2=2)
-        model = introduce_vacancies(model, 0.2)
-    
-        motif_hash = hash_numpy_array(model.motif)
-        
-        assert math.isclose(motif_hash, 171.9359264901)
+    assert math.isclose(motif_hash, 171.9359264901)
     
 
 def test_impurities():
@@ -57,17 +55,16 @@ def test_impurities():
     
     np.random.seed(1)
     
-    path = INPUT_FILE_DIR / "hBN.txt"
-    with open(path) as fp:
-        config = parse_config_file(fp)
+    path = INPUT_FILE_DIR / "hBN.yaml"
+    config = parse_config_file(path)
+
+    model = SlaterKoster(config).supercell(n1=2)
+    model.initialize_hamiltonian()
+
+    model = introduce_impurities(model, 10, 0.9)
+
+    hamiltonian_hash = np.real(hash_numpy_array(model.hamiltonian[0]))
     
-        model = SlaterKoster(config).supercell(n1=2)
-        model.initialize_hamiltonian()
-    
-        model = introduce_impurities(model, 10, 0.9)
-    
-        hamiltonian_hash = np.real(hash_numpy_array(model.hamiltonian[0]))
-        
-        assert math.isclose(hamiltonian_hash, 431.73999999999995)
+    assert math.isclose(hamiltonian_hash, 431.73999999999995)
     
     

[…]
