### 2023-11-16 af29a15
Remove old input files
