+++ examples/inputs/germanene.yaml
@@ -1,0 +1,19 @@
+SystemName: Germanene
+Dimensions: 2
+Lattice: 
+  - [3.48142212,  2.01, 0.0]
+  - [3.48142212, -2.01, 0.0]
+Species: Ge
+Motif:
+  - [0, 0, 0, 0]
+  - [2.3209481, 0, -0.687496, 0]
+Filling: 4
+Orbitals: [s px py pz]
+OnsiteEnergy:
+  - [-7.9, -1.16, -1.16, -1.16]
+SKAmplitudes: 
+  - (0, 0) -1.79 2.36 4.15 -1.04
+Spin: True
+SOC: 0.294
+Mesh: [200, 200]
+SymmetryPoints: G M K G
