+++ examples/inputs/MoS2_Ridolfi.yaml
@@ -1,0 +1,26 @@
+SystemName: MoS2
+Dimensions: 2
+Lattice: 
+  - [3.16, 0.0, 0.0]
+  - [1.58, 2.73664, 0.0]
+Species: [Mo, S]
+Motif:
+  - [0, 0, 0, 0]
+  - [0.0, 1.8246186736331629,  1.5683120530778172, 1]
+  - [0.0, 1.8246186736331629, -1.5683120530778172, 1]
+Filling: [6, 4]
+Orbitals: 
+  - dxy dyz dzx dx2-y2 d3z2-r2
+  - px py pz
+OnsiteEnergy:
+  - [-0.352, -1.563, -1.563, -0.352, 0.201]
+  - [-54.839, -54.839, -39.275]
+SKAmplitudes: 
+  - (0, 0; 1) -1.153 0.612 0.086
+  - (0, 1; 1) -9.880 4.196
+  - (1, 1; 1) 12.734 -2.175
+Spin: True
+SOC: [0.1125, 0.078]
+Radius: 3.16
+Mesh: [200, 200]
+SymmetryPoints: G K M G
