### 2023-11-16 170f399
Update example calculations
