+++ examples/amorphous_slater_koster.py
@@ -7,6 +7,7 @@ from tightbinder.observables import dos
 from tightbinder.fileparse import parse_config_file
 import matplotlib.pyplot as plt
 import numpy as np
+from pathlib import Path
 
 
 def main():
@@ -18,8 +19,8 @@ def main():
     disorder = 0.1
 
     # Parse configuration file
-    file = open("./examples/Bi111.txt", "r")
-    config = parse_config_file(file)
+    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.yaml"
+    config = parse_config_file(path)
 
     # Init. model and construct supercell
     first_neighbour_distance = np.linalg.norm(config["Motif"][1][:3])
