+++ examples/bands_ribbon_bi111.py
@@ -1,12 +1,13 @@
 from tightbinder.models import SlaterKoster
 from tightbinder.fileparse import parse_config_file
 import matplotlib.pyplot as plt
+from pathlib import Path
 
 def main():
 
     # Parse configuration file
-    file = open("./examples/Bi111.txt", "r")
-    config = parse_config_file(file)
+    path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.yaml"
+    config = parse_config_file(path)
 
     # Init. model and consider a finite supercell along one Bravais vector
     width = 15
