### 2023-11-16 b104067
Update docs to reflect new input format
