+++ docs/source/materials/materials/mos2.rst
@@ -22,92 +22,32 @@ The configuration file is:
 .. code-block::
     :caption: examples/MoS2_Ridolfi.txt
 
-    ! --------------------- Config file formatting rules ---------------------
-    ! Lines starting with ! are comments
-    ! Lines with # denote argument types and must be present (removing one or wirting it wrogn will result in error)
-    ! Blank spaces are irrelevant
-    ! All numbers must be separated by either spaces, commas or semicolons
-    ! Beware: Leaving a blank space at the beginning of a line will result in it being skipped
-    ! -------------------------- Material parameters -------------------------
-    #System name
-    ! Model taken from doi:10.1088/0953-8984/27/36/365501
-    MoS2
-
-    #Dimensionality
-    2
-
-    #Bravais lattice
-    ! Each row corresponds to one vector of the basis (vectors have to be in R^3)
-    ! For 1D and 2D systems, the z (third) component must be zero (z should always be regarded as height component)
-    3.16 0.0 0.0
-    1.58 2.73664 0.0
-
-    #Species
-    ! Number of atomic species.
-    2
-
-    #Motif
-    ! Each row corresponds to one vector of the basis; fourth number denotes atomic species
-    ! Numbering of species starts at 0.
-    0 0 0 0
-    0.0 1.8246186736331629 1.5683120530778172 1
-    0.0 1.8246186736331629 -1.5683120530778172 1
-
-    #Orbitals
-    ! Always use cubic harmonic orbitals
-    ! Order is irrelevant
-    ! Possible orbitals are: s, px, py, pz, dxy, dyz, dzx, dx2-y2, d3z2-r2
-    dxy dyz dzx dx2-y2 d3z2-r2
-    px py pz
-
-    #Filling
-    ! Specify number of electrons per chemical species; one line per species. 
-    ! Fillings can be fractional, but total number of electons in unit cell must be an integer.
-    6
-    4
-
-    #Onsite energy
-    ! One for each different orbital defined above.
-    ! In case of having several atomic species multiple line must be specified, one for each element.
-    -0.352 -1.563 -1.563 -0.352 0.201
-    -54.839 -54.839 -39.275
-
-    #SK amplitudes
-    ! Order is fixed: Vsss, Vsps, Vpps, Vppp, Vsds, Vpds, Vpdp, Vdds, Vddp, Vddd (separated by spaces).
-    ! Not present amplitudes can be omitted but ordering has te be respected always.
-    ! The syntax for SK amplitudes allows to specify between which species are the hoppings, as well as
-    ! which is the corresponding neighbour. 
-    ! Syntax is: [species1, species2; n-th neighbour] V1 V2 ...
-    [0, 0; 1] -1.153 0.612 0.086
-    [0, 1; 1] -9.880 4.196
-    [1, 1; 1] 12.734 -2.175
-
-    #Spin
-    ! Determine whether the model is spinless (spin polarized) or spinful.
-    ! Must be either True (spinful) or False (spinless); if left blank it will default to False.
-    ! NOTE: True or False must be capitalized.
-    True
-
-    #Spin-orbit coupling
-    ! Note: Using a non-zero value will automatically produce a spinful model.
-    ! Amplitude must be specified for all species; one line per species.
-    0.1125
-    0.078
-
-    ! --------------------- Simulation parameters ---------------------
-    #Radius
-    ! If present, SlaterKoster model runs in "radius" mode, meaning that it will look
[…]
