+++ examples/inputs/chain.yaml
@@ -8,7 +8,7 @@ Motif:
 Filling: 1
 Orbitals: s
 OnsiteEnergy:
-  - [1]
+  - 1
 SKAmplitudes: 
   - (0, 0) -0.3
 Spin: True

[…]
