+++ docs/source/config.rst
@@ -7,7 +7,7 @@ different chemical species present, the orbitals for each different type of atom
 of valid configuration files can be found under the folder ``/examples`` in the repository, and are also present in the :doc:`Materials <../materials/index>` section. 
 Here we go over the different fields expected in the configuration files, as well as the syntaxis used.
 
-The configuration files are defined using the ![YAML](https://en.wikipedia.org/wiki/YAML) format. This format allows to write the configuration files with a structure
+The configuration files are defined using the `YAML`<https://en.wikipedia.org/wiki/YAML>_ format. This format allows to write the configuration files with a structure
 matching the interal storage of the information (this is, a dictionary), also in a readable way for humans. 
 Each field is defined by a keyword, followed by a colon and the corresponding value(s).
 When a field denotes some kind of list, there are two forms to write it: either by enclosing the elements in square brackets and separating them by commas, or by
@@ -107,9 +107,9 @@ NOTE: True or False must be capitalized to follow Python naming.
 
 .. code-block:: 
     
-    [Spin-orbit coupling]
+    [SOC]
 
-Using a non-zero value of spin-orbit coupling will automatically produce a spinful model, i.e. it sets 
+Field for spin-orbit coupling. Using a non-zero value of spin-orbit coupling will automatically produce a spinful model, i.e. it sets 
 the ``spin`` option to ``True`` even if it was set to ``False`` before. The amplitudes must be specified for all species, one value per species following the YAML syntax
 for lists.
 If there is only one value, it can be written without brackets.
