### 2023-11-16 a1f0c5a
Updated example scripts in docs
