+++ docs/source/examples/scripts/amorphous_slater_koster.rst
@@ -19,6 +19,7 @@ Here we showcase this class, and use it to compare the density of states of the 
     from tightbinder.fileparse import parse_config_file
     import matplotlib.pyplot as plt
     import numpy as np
+    from pathlib import Path
 
 
     def main():
@@ -30,8 +31,8 @@ Here we showcase this class, and use it to compare the density of states of the 
         disorder = 0.1
 
         # Parse configuration file
-        file = open("./examples/Bi111.txt", "r")
-        config = parse_config_file(file)
+        path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.yaml"
+        config = parse_config_file(path)
 
         # Init. model and construct supercell
         first_neighbour_distance = np.linalg.norm(config["Motif"][1][:3])
@@ -52,7 +53,6 @@ Here we showcase this class, and use it to compare the density of states of the 
         model.initialize_hamiltonian()
         results_amorphous = model.solve()
         results_amorphous.rescale_bands_to_fermi_level()
-        
         dos_amorphous, energies_amorphous = dos(results_amorphous, delta=delta, npoints=dos_npoints)
 
         # Plot both densities of states
