+++ docs/source/examples/scripts/bands_ribbon_bi111.rst
@@ -12,12 +12,13 @@ which is specific to honeycomb lattices.
     from tightbinder.models import SlaterKoster
     from tightbinder.fileparse import parse_config_file
     import matplotlib.pyplot as plt
+    from pathlib import Path
 
     def main():
 
         # Parse configuration file
-        file = open("./examples/Bi111.txt", "r")
-        config = parse_config_file(file)
+        path = Path(__file__).parent / ".." / "examples" / "inputs" / "Bi111.yaml"
+        config = parse_config_file(path)
 
         # Init. model and consider a finite supercell along one Bravais vector
         width = 15
