+++ docs/source/examples/scripts/density_states_kpm.rst
@@ -13,6 +13,7 @@ the DoS without having to diagonalize previously the Hamiltonian, which is the s
     from tightbinder.observables import dos_kpm
     import matplotlib.pyplot as plt
     import numpy as np
+    from pathlib import Path
 
     def main():
         
@@ -20,8 +21,8 @@ the DoS without having to diagonalize previously the Hamiltonian, which is the s
         np.random.seed(1)
 
         # Parse configuration file and init. model
-        file = open("./examples/hBN.txt", "r")
-        config = parse_config_file(file)
+        path = Path(__file__).parent / ".." / "examples" / "inputs" / "hBN.yaml"
+        config = parse_config_file(path)
 
         ncells = 25
         model = SlaterKoster(config).supercell(n1=ncells, n2=ncells)
