+++ docs/source/examples/scripts/transport_hydrogen_nanotube.rst
@@ -10,6 +10,7 @@ Transmission of a nanotube
     import matplotlib.pyplot as plt
     from matplotlib.gridspec import GridSpec
     import numpy as np
+    from pathlib import Path
 
     def main():
 
@@ -21,8 +22,8 @@ Transmission of a nanotube
         ax3d = plt.figure().add_subplot(projection='3d')
 
         # Parse configuration file
-        file = open("examples/chain.txt", "r")
-        config = parse_config_file(file)
+        path = Path(__file__).parent / ".." / "examples" / "inputs" / "chain.yaml"
+        config = parse_config_file(path)
 
         # Init. model
         model = SlaterKoster(config)
