### 2024-01-11 #5810 @ppxasjsm
Hi @alejandrojuria could you work on this list of things please:

I'll go through the paper an may make some small editorial suggestions either as a PR or a list of small fixes here.
