+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 
 name = "tightbinder"  
-version = "0.2"  
+version = "0.2.0"  
 description = "Slater-Koster tight-binding library for crystalline and disordered solids"  
 readme = "README.md" 
 requires-python = ">=3.7"
