### 2024-01-12 #5810 @alejandrojuria
Hi @ppxasjsm, I believe I'm done with the tasks above for the author (although I seem unable to check the boxes). Current release version is v0.2.1, and the DOI for the Zenodo release is [10.5281/zenodo.10496449](https://zenodo.org/doi/10.5281/zenodo.10496449)
